changes the BCLK config

+++ SAMV71x/app/12_Fft - Afec/src/Bsw/Ecual/CODEC/include/CODEC.h
@@ -19,5 +19,5 @@
 void CODEC_Init( void );
 void CODEC_StartAudioCapture_MONO( void );
 void CODEC_StopAudioCapture_MONO( void );
-void CODEC_PrintAudioCaptured_MONO(int16_t *data, uint32_t size);
+void CODEC_PrintAudioCaptured_MONO(uint32_t *data, uint32_t size);
 #endif
